+++ audio_service/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Add AudioServiceFragmentActivity.
 * Support `content://` art URIs in notification on Android (@nt4f04uNd)
+* Document Android foregroundServiceType.
 
 ## 0.18.4